third_party/WebKit/LayoutTests/inspector/console/console-log-object-with-getter-expected.txt

CONSOLE MESSAGE: line 11: [object Object]
Tests that console logging dumps object values defined by getters and allows to expand it.

console-log-object-with-getter.html:11 Objectfoo: Objecta: 1b: 2__proto__: Objectset bar: function (x)get foo: function ()__proto__: Object
